feat: ADD message change logic, fix data structure 싸피마인드 메세지 교환 로직추가, 자료구조 수정

+++ backend/src/main/java/com/playssafy/playssafy/dto/ssafymind/Point.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 public class Point {
     private int x; // x좌표
     private int y; // y좌표
-    private int size; // 선 굵기
+    private double size; // 선 굵기
     private String color; // 색
     private boolean fillFlag; // 채우기 여부
 }
